requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
Werkzeug==3.0.1
gunicorn==21.2.0
flask_sqlalchemy== 26.0.1